Various ABDK audit fixes (#785)
* CVF 92

* forge fmt

* Reverting CVF92

* CVF 24, 25, 26

* cvf 27

* CVF 29, 30

* comment for CVF 39

* cvf 161, 162, 163

* CVF 143

* CVF 141

* CVF 157 158

* rearranging to improve gas instead of cvf 116

* Protocol fee return explained

* fix compilation

* cvf 82

* CVF 42

* CVF 56

* cvf 40

* cvf 61

* CVF 62

* Update src/interfaces/IPoolManager.sol


* PR comments

* Extend fuzz test

---------

+++ src/libraries/ProtocolFeeLibrary.sol
@@ -32,7 +32,8 @@ library ProtocolFeeLibrary {
     // The protocol fee is taken from the input amount first and then the LP fee is taken from the remaining
     // The swap fee is capped at 100%
     // Equivalent to protocolFee + lpFee(1_000_000 - protocolFee) / 1_000_000
-    function calculateSwapFee(uint24 self, uint24 lpFee) internal pure returns (uint24 swapFee) {
+    /// @dev here `self` is just a single direction's protocol fee, not a packed type of 2 protocol fees
+    function calculateSwapFee(uint16 self, uint24 lpFee) internal pure returns (uint24 swapFee) {
         // protocolFee + lpFee - (protocolFee * lpFee / 1_000_000). Div rounds up to favor LPs over the protocol.
         assembly {
             let numerator := mul(self, lpFee)
